Combine Timeline components
